Efficiency and Missing points
Fix situation where no tracking points in first sector. If we have points in or close to any other sector then we will assume that the start we recorded before is valid. Sadly means checking each sector which is crap. To help make up for that don't recheck start times as often if the last start was more than 30 minutes ago