General performance improvements (#15)
* add pm2 configuration file

* Only detect launching within 30 km

* Match on registration properly

Ignore unwanted characters in ddb registration field. These are stripped from the soaringspot field so must be stripped here.
Also allow either registration or compno to be defined

* Don't fail if channel doesn't exist but has viewers

* pm2 fixes

- performance counters for OGN
- start .js directly rather than using yarn

* Change requested terrain tile resolution

- now approximately 1.8m per pixel which means a lot less tiles will be loaded

* Don't expire tiles that are being accessed

* Auto display altitude fix

If speed or distance are too low then we will remain on height. This mostly covers situation when gliders have started but don't actually start

* Don't calculate vario if no points

* Display ! icon when no tracking

* Handle missing tasks better