Correct AGL handling (#16)
* yarn.lock is local

* Remove origin as appears to now break mapbox

* Correct min/max altitude display

* Load lat/lng from the competition table in the database

* Shortcut to tracker rather than repeated lookup

* Ensure glider and tracker point to the same object

Previously this was creating a copy, and only tracker was getting the data. This change ensures that both glider and tracker point to the same object which means scores get correctly enriched with height and vario information

* Use gliderKey rather than mergedName

* Round AGL to 0 decimal points and save in glider

Without saving it doesn't get mixed in properly which meant that height AGL would sometimes vanish

* Ensure table names are the correct case

If mysql is case sensitive for table names then these queries fail. Thanks to @lc1lc

Co-Authored-By: lc1lc <12581685+lc1lc@users.noreply.github.com>

* Warning message for when TZ table not populated

Thanks @lc1lc

Co-Authored-By: lc1lc <12581685+lc1lc@users.noreply.github.com>

Co-authored-by: lc1lc <12581685+lc1lc@users.noreply.github.com>